update: split patching
