+++ src/py/mat3ra/notebooks_utils/pyodide/packages/torch.py
@@ -1,24 +1,13 @@
 """
 Patches for PyTorch and related packages to work in Pyodide environment.
 
-This module provides patches for various torch-related packages that don't work
-in Pyodide's WASM environment, organized by functionality.
+This module provides shared Pyodide compatibility patches for torch-based MLFF
+packages. Package-specific patches live in their corresponding package modules.
 
 Usage:
-    from mat3ra.notebooks_utils.other.torch_pyodide import (
-        patch_torch_linalg,
-        patch_torch_testing,
-        patch_matscipy,
-        patch_lmdb_h5py,
-        patch_mace_tools,
-    )
+    from mat3ra.notebooks_utils.pyodide.packages import apply_all_patches
 
-    # Apply all patches
-    patch_torch_linalg()
-    patch_torch_testing()
-    patch_matscipy()
-    patch_lmdb_h5py()
-    patch_mace_tools()
+    apply_all_patches("mace")
 """
 
 import sys
@@ -32,6 +21,49 @@ import torch
 EigRet = namedtuple("linalg_eig", ["eigenvalues", "eigenvectors"])  # type: ignore
 EighRet = namedtuple("linalg_eigh", ["eigenvalues", "eigenvectors"])  # type: ignore
 LUFactorReturn = namedtuple("LUFactorReturn", ["LU", "pivots"])
+_OMEGACONF_RESOLVERS = {}
+
+
+class _DictConfig(dict):
+    pass
+
+
+class _ListConfig(list):
+    pass
+
+
+class _OmegaConf:
+    @staticmethod
+    def to_container(cfg, **kwargs):
+        return dict(cfg) if isinstance(cfg, dict) else cfg
+
+    @staticmethod
+    def create(data):
+        return _DictConfig(data) if isinstance(data, dict) else data
+
+    @staticmethod
+    def register_new_resolver(name, func, **kwargs):
+        _OMEGACONF_RESOLVERS[name] = func
+
+    @staticmethod
+    def register_resolver(name, func, **kwargs):
+        _OMEGACONF_RESOLVERS[name] = func
+
+    @staticmethod
+    def has_resolver(name):
+        return name in _OMEGACONF_RESOLVERS
+
+
+def _ensure_omegaconf_stub():
+    omegaconf_mod = sys.modules.get("omegaconf") or _make_stub_module("omegaconf")
+    omegaconf_mod.DictConfig = getattr(omegaconf_mod, "DictConfig", _DictConfig)
+    omegaconf_mod.ListConfig = getattr(omegaconf_mod, "ListConfig", _ListConfig)
+    omega_conf = getattr(omegaconf_mod, "OmegaConf", _OmegaConf)
+    for name in ("register_new_resolver", "register_resolver", "has_resolver"):
+        if not hasattr(omega_conf, name):
+            setattr(omega_conf, name, getattr(_OmegaConf, name))
+    omegaconf_mod.OmegaConf = omega_conf
+    return omegaconf_mod
 
 
 def _to_np(tensor):
@@ -543,7 +575,7 @@ def patch_torch_distributed():
 
 
 # ==============================================================================
-# FAIRChem heavy dependency stubs
+# Stub helpers
 # ==============================================================================
 
 
@@ -570,313 +602,6 @@ def _make_stub_module(name, attrs=None, submodules=None):
             setattr(mod, sub_name, sub_mod)
             sys.modules[full_name] = sub_mod
     return mod
-
-
-def patch_fairchem_deps():
-    """
-    Stub heavy dependencies that fairchem-core imports but doesn't need for inference.
-
-    This stubs: numba, ray (+ serve), wandb, torchtnt, hydra, omegaconf,
-    submitit, clusterscope, tqdm, huggingface_hub, websockets.
-    """
-    # --- numba ---
-    numba_mod = _make_stub_module("numba", submodules=["core", "core.types", "typed"])
-    numba_mod.njit = lambda *a, **k: (lambda f: f) if not a or callable(a[0]) else lambda f: f
-    numba_mod.jit = numba_mod.njit
-    numba_mod.prange = range
-    for t in ("int32", "int64", "float32", "float64", "boolean"):
-        setattr(numba_mod, t, t)
-
-    class _TypedList(list):
-        pass
-
-    sys.modules["numba.typed"].List = _TypedList
-
-    # --- ray ---
-    ray_mod = _make_stub_module(
-        "ray",
-        submodules=[
-            "serve",
-            "runtime_env",
-            "train",
-            "data",
-            "util",
-            "util.scheduling_strategies",
-            "util.queue",
-        ],
-    )
-
-    def _ray_remote(*args, **kwargs):
-        if args and callable(args[0]):
-            args[0]._remote = lambda *a, **kw: None
-            return args[0]
-
-        def wrapper(fn_or_cls):
-            fn_or_cls._remote = lambda *a, **kw: None
-            return fn_or_cls
-
-        return wrapper
-
-    ray_mod.remote = _ray_remote
-    ray_mod.init = lambda *a, **k: None
-    ray_mod.get = lambda *a, **k: None
-    ray_mod.put = lambda *a, **k: None
-    ray_mod.wait = lambda *a, **k: ([], [])
-    ray_mod.is_initialized = lambda: False
-
-    class _ObjectRef:
-        pass
-
-    ray_mod.ObjectRef = _ObjectRef
-
-    class _PlacementGroupSchedulingStrategy:
-        def __init__(self, *a, **k):
-            pass
-
-    sys.modules["ray.util.scheduling_strategies"].PlacementGroupSchedulingStrategy = _PlacementGroupSchedulingStrategy
-
-    # ray.serve stubs
-    _serve = sys.modules["ray.serve"]
-
-    def _serve_deployment(*args, **kwargs):
-        if args and callable(args[0]):
-            return args[0]
-        return lambda cls_or_fn: cls_or_fn
-
-    _serve.deployment = _serve_deployment
-    _serve.ingress = lambda *a, **k: (lambda cls: cls)
-    _serve.run = lambda *a, **k: None
-    _serve.batch = lambda *args, **kwargs: (lambda fn: fn) if not args or not callable(args[0]) else args[0]
-
-    _serve_schema = types.ModuleType("ray.serve.schema")
-    _serve_schema.__package__ = "ray.serve"
-
-    class _LoggingConfig:
-        def __init__(self, **k):
-            self.__dict__.update(k)
-
-    _serve_schema.LoggingConfig = _LoggingConfig
-    _serve.schema = _serve_schema
-    sys.modules["ray.serve.schema"] = _serve_schema
-
-    # --- wandb ---
-    wandb_mod = _make_stub_module("wandb")
-    wandb_mod.init = lambda *a, **k: None
-    wandb_mod.log = lambda *a, **k: None
-    wandb_mod.finish = lambda *a, **k: None
-
-    # --- torchtnt ---
-    _make_stub_module(
-        "torchtnt",
-        submodules=[
-            "framework",
-            "framework.state",
-            "framework.unit",
-            "framework.callback",
-            "framework.auto_unit",
-            "framework.fit",
-            "framework.train",
-            "framework.evaluate",
-            "framework.predict",
-            "utils",
-            "utils.loggers",
-            "utils.timer",
-            "utils.distributed",
-            "utils.prepare_module",
-        ],
-    )
-
-    class _PredictUnit:
-        def __init__(self, *a, **k):
-            pass
-
-        def __class_getitem__(cls, item):
-            return cls
-
-    class _TrainUnit:
-        def __init__(self, *a, **k):
-            pass
-
-        def __class_getitem__(cls, item):
-            return cls
-
-    class _EvalUnit:
-        def __init__(self, *a, **k):
-            pass
-
-        def __class_getitem__(cls, item):
-            return cls
-
-    class _State:
-        def __init__(self, *a, **k):
-            pass
-
-        def __class_getitem__(cls, item):
-            return cls
-
-    class _Callback:
-        pass
-
-    tnt_framework = sys.modules["torchtnt.framework"]
-    tnt_unit = sys.modules["torchtnt.framework.unit"]
-    tnt_state = sys.modules["torchtnt.framework.state"]
-    tnt_cb = sys.modules["torchtnt.framework.callback"]
-    for m in [tnt_framework, tnt_unit]:
-        m.PredictUnit = _PredictUnit
-        m.TrainUnit = _TrainUnit
-        m.EvalUnit = _EvalUnit
-    tnt_state.State = _State
-    tnt_cb.Callback = _Callback
-    tnt_framework.State = _State
-    tnt_framework.Callback = _Callback
-
-    # torchtnt entry point functions
-    sys.modules["torchtnt.framework.fit"].fit = lambda *a, **k: None
-    sys.modules["torchtnt.framework.train"].train = lambda *a, **k: None
-    sys.modules["torchtnt.framework.evaluate"].evaluate = lambda *a, **k: None
-    sys.modules["torchtnt.framework.predict"].predict = lambda *a, **k: None
-
-    # torchtnt.utils stubs
-    tnt_dist = sys.modules["torchtnt.utils.distributed"]
-    tnt_dist.get_file_init_method = lambda *a, **k: ""
-    tnt_dist.get_tcp_init_method = lambda *a, **k: ""
-    tnt_dist.spawn_multi_process = lambda *a, **k: None
-
-    tnt_prep = sys.modules["torchtnt.utils.prepare_module"]
-    tnt_prep.prepare_module = lambda module, *a, **k: module
-    tnt_prep.FSDPStrategy = type("FSDPStrategy", (), {"__init__": lambda self, **k: None})
-    tnt_prep.DDPStrategy = type("DDPStrategy", (), {"__init__": lambda self, **k: None})
-    tnt_prep.NOOPStrategy = type("NOOPStrategy", (), {"__init__": lambda self, **k: None})
-
-    # --- hydra / omegaconf ---
-    omegaconf_mod = _make_stub_module("omegaconf")
-
-    class _DictConfig(dict):
-        pass
-
-    class _ListConfig(list):
-        pass
-
-    omegaconf_mod.DictConfig = _DictConfig
-    omegaconf_mod.ListConfig = _ListConfig
-    omegaconf_mod.OmegaConf = type(
-        "OmegaConf",
-        (),
-        {
-            "to_container": staticmethod(lambda cfg, **k: dict(cfg) if isinstance(cfg, dict) else cfg),
-            "create": staticmethod(lambda d: _DictConfig(d) if isinstance(d, dict) else d),
-        },
-    )
-    _make_stub_module("hydra", submodules=["core", "core.global_hydra", "utils"])
-
-    def _hydra_instantiate(config, *args, _recursive_=True, **kwargs):
-        import importlib
-
-        if isinstance(config, dict) and "_target_" in config:
-            target = config["_target_"]
-            mod_path, cls_name = target.rsplit(".", 1)
-            mod = importlib.import_module(mod_path)
-            cls = getattr(mod, cls_name)
-            pos_args = list(args) + list(config.get("_args_", []))
-            cfg = {k: v for k, v in config.items() if not k.startswith("_")}
-            if _recursive_:
-                for k, v in cfg.items():
-                    if isinstance(v, dict) and "_target_" in v:
-                        cfg[k] = _hydra_instantiate(v)
-                    elif isinstance(v, list):
-                        cfg[k] = [_hydra_instantiate(i) if isinstance(i, dict) and "_target_" in i else i for i in v]
-                pos_args = [_hydra_instantiate(a) if isinstance(a, dict) and "_target_" in a else a for a in pos_args]
-            return cls(*pos_args, **{**cfg, **kwargs})
-        return config
-
-    sys.modules["hydra.utils"].instantiate = _hydra_instantiate
-    sys.modules["hydra"].utils = sys.modules["hydra.utils"]
-
-    # --- submitit / clusterscope ---
-    _make_stub_module("submitit")
-    _make_stub_module("clusterscope")
-
-    # --- websockets ---
-    _make_stub_module("websockets")
-
-    # --- tqdm ---
-    tqdm_mod = _make_stub_module("tqdm", submodules=["auto", "std"])
-
-    def _tqdm_passthrough(iterable=None, *a, **k):
-        return iterable if iterable is not None else iter([])
-
-    tqdm_mod.tqdm = _tqdm_passthrough
-    sys.modules["tqdm.auto"].tqdm = _tqdm_passthrough
-    sys.modules["tqdm.std"].tqdm = _tqdm_passthrough
-
-    # --- huggingface_hub ---
-    hf_mod = _make_stub_module("huggingface_hub", submodules=["utils"])
-    hf_mod.hf_hub_download = lambda *a, **k: ""
-    hf_mod.snapshot_download = lambda *a, **k: ""
-
-    # --- ase_db_backends ---
-    _make_stub_module("ase_db_backends")
-    # --- INT8 quantized model support ---
-    _orig_torch_load = torch.load
-
-    def _int8_aware_torch_load(f, *args, **kwargs):
-        result = _orig_torch_load(f, *args, **kwargs)
-        if isinstance(result, dict) and "quantized_ema_state_dict" in result:
-            import gc as _gc
-
-            from fairchem.core.units.mlip_unit.api.inference import MLIPInferenceCheckpoint
-
-            print("  Dequantizing INT8 -> FP16 (streaming)...")
-            quantized_ema = result.pop("quantized_ema_state_dict")
-            scales = result.pop("quantization_scales")
-            ema_state_dict = {}
-            names = list(quantized_ema.keys())
-            for name in names:
-                tensor = quantized_ema.pop(name)
-                if name in scales:
-                    scale = scales.pop(name)
-                    ema_state_dict[name] = (tensor.float() * scale.float()).half()
-                    del scale
-                else:
-                    ema_state_dict[name] = tensor
-                del tensor
-            del quantized_ema, scales
-            _gc.collect()
-            checkpoint = MLIPInferenceCheckpoint(
-                model_config=result["model_config"],
-                model_state_dict=result.get("model_state_dict", {}),
-                ema_state_dict=ema_state_dict,
-                tasks_config=result["tasks_config"],
-            )
-            del result
-            _gc.collect()
-            print("  ✓ Dequantization complete")
-            return checkpoint
-        return result
-
-    torch.load = _int8_aware_torch_load
-
-    # --- Model registry fallback ---
-    try:
-        from fairchem.core.common.registry import registry as _registry
-
-        _orig_get_model = _registry.get_model_class
-
-        def _fallback_get_model_class(name):
-            try:
-                return _orig_get_model(name)
-            except RuntimeError:
-                import importlib as _imp
-
-                module_path, class_name = name.rsplit(".", 1)
-                mod = _imp.import_module(module_path)
-                return getattr(mod, class_name)
-
-        _registry.get_model_class = _fallback_get_model_class
-    except Exception:
-        pass
-
-    print("✓ FAIRChem dependency stubs applied")
 
 
 # ==============================================================================
@@ -913,604 +638,9 @@ def patch_matscipy():
     print("✓ Matscipy patches applied")
 
 
-# ==============================================================================
-# LMDB and HDF5 patches
-# ==============================================================================
-
-
-def patch_mace_training():
-    """
-    Stub lmdb and h5py packages.
-
-    These are C-extension packages used by MACE's training/dataset code
-    but not needed for inference. Stubs allow imports to succeed.
-    """
-    for _pkg in ("lmdb", "h5py"):
-        if _pkg not in sys.modules:
-            sys.modules[_pkg] = types.ModuleType(_pkg)
-
-    print("✓ LMDB and HDF5 stubs applied")
-
-
-# ==============================================================================
-# MACE tools patches
-# ==============================================================================
-
-
-def patch_mace_tools():
-    """
-    Fix MACE's torch_geometric import order issues in Pyodide.
-
-    In Pyodide, torch_geometric.data may not be set during circular imports.
-    Pre-importing ensures the attribute is available when MACE needs it.
-    """
-    try:
-        import importlib as _importlib
-
-        _tg = _importlib.import_module("mace.tools.torch_geometric")
-        _tg_data = _importlib.import_module("mace.tools.torch_geometric.data")
-        _tg.data = _tg_data
-        print("✓ MACE tools patches applied")
-    except Exception as e:
-        print(f"⚠ MACE tools patches skipped: {e}")
-
-
-# ==============================================================================
-# Convenience function to apply all patches
-# ==============================================================================
-
-
-def apply_all_patches(include_fairchem=False, include_mattersim=False, include_nequip=False):
-    """
-    Apply all torch and model patches for Pyodide in one call.
-
-    Args:
-        include_fairchem: If True, also apply FAIRChem-specific patches
-            (torch.distributed, heavy dependency stubs). Set this when
-            using fairchem-core / UMA models.
-        include_mattersim: If True, also apply MatterSim-specific patches
-            (loguru, azure, e3nn JIT stubs). Set this when
-            using MatterSim / M3GNet models.
-        include_nequip: If True, also apply NequIP-specific patches
-            (lightning, hydra, torchmetrics, e3nn JIT stubs). Set this
-            when using NequIP models.
-    """
+def apply_common_torch_patches():
+    """Apply shared Pyodide patches needed by torch-based MLFF packages."""
     patch_torch_linalg()
     patch_torch_compiler()
     patch_torch_testing()
     patch_matscipy()
-    patch_mace_training()
-    patch_mace_tools()
-
-    if include_fairchem:
-        patch_torch_distributed()
-        patch_fairchem_deps()
-
-    if include_mattersim:
-        patch_torch_distributed()
-        patch_mattersim_deps()
-
-    if include_nequip:
-        patch_nequip_deps()
-
-    print("\n✅ All Pyodide patches applied successfully!")
-
-
-# ==============================================================================
-# MatterSim patches
-# ==============================================================================
-
-
-def patch_mattersim_deps():
-    """
-    Stub heavy dependencies required by MatterSim but not needed for inference.
-
-    Stubs: loguru, azure.*, atomate2, seekpath, phonopy, phono3py, mp_api,
-    sklearn, and patches e3nn to disable JIT/torch.compile.
-    """
-    loguru_mod = _make_stub_module("loguru")
-
-    class _Logger:
-        def info(self, msg, *a, **k):
-            print(f"INFO: {msg}")
-
-        def warning(self, msg, *a, **k):
-            print(f"WARNING: {msg}")
-
-        def error(self, msg, *a, **k):
-            print(f"ERROR: {msg}")
-
-        def debug(self, msg, *a, **k):
-            pass
-
-        def trace(self, msg, *a, **k):
-            pass
-
-        def success(self, msg, *a, **k):
-            print(f"✓ {msg}")
-
-        def __getattr__(self, name):
-            return lambda *a, **k: None
-
-    loguru_mod.logger = _Logger()
-
-    _make_stub_module("azure", submodules=["identity", "storage", "storage.blob"])
-
-    for pkg in [
-        "atomate2",
-        "seekpath",
-        "phonopy",
-        "phono3py",
-        "mp_api",
-        "jobflow",
-        "emmet",
-        "emmet.core",
-        "emmet.core.tasks",
-        "maggma",
-    ]:
-        _make_stub_module(pkg)
-
-    _make_stub_module(
-        "sklearn",
-        submodules=[
-            "base",
-            "utils",
-            "utils.validation",
-            "preprocessing",
-            "model_selection",
-            "gaussian_process",
-            "gaussian_process.kernels",
-        ],
-    )
-
-    class _GPR:
-        def __init__(self, *a, **k):
-            pass
-
-        def fit(self, *a, **k):
-            return self
-
-        def predict(self, X, return_std=False):
-            import numpy as _np
-
-            mean = _np.zeros(X.shape[0])
-            if return_std:
-                return mean, _np.ones(X.shape[0])
-            return mean
-
-        def log_marginal_likelihood(self):
-            return 0.0
-
-    sys.modules["sklearn.gaussian_process"].GaussianProcessRegressor = _GPR
-
-    class _Kernel:
-        pass
-
-    class _DotProduct(_Kernel):
-        def __init__(self, *a, **k):
-            pass
-
-    class _Hyperparameter:
-        def __init__(self, *a, **k):
-            pass
-
-    _sk_kernels = sys.modules["sklearn.gaussian_process.kernels"]
-    _sk_kernels.Kernel = _Kernel
-    _sk_kernels.DotProduct = _DotProduct
-    _sk_kernels.Hyperparameter = _Hyperparameter
-
-    try:
-        import pyodide_http  # noqa: F401
-
-        pyodide_http.patch_all()
-    except ImportError:
-        pass
-
-    try:
-        import e3nn
-
-        e3nn._SO3_INITIALIZED = True
-    except Exception:
-        pass
-
-    import torch
-
-    if not hasattr(torch.jit, "_original_script"):
-        _orig_script = torch.jit.script
-
-        def _noop_script(obj=None, *a, **k):
-            if obj is not None:
-                return obj
-            return lambda fn: fn
-
-        torch.jit.script = _noop_script
-
-    _te = _make_stub_module("torch_ema")
-
-    class _EMA:
-        def __init__(self, *a, **k):
-            pass
-
-    _te.ExponentialMovingAverage = _EMA
-
-    _tm = _make_stub_module("torchmetrics")
-
-    class _MeanMetric:
-        def __init__(self, *a, **k):
-            pass
-
-    _tm.MeanMetric = _MeanMetric
-
-    _make_stub_module("torch_geometric", submodules=["data", "loader", "utils"])
-
-    class _Data:
-        def __init__(self, **kwargs):
-            for k, v in kwargs.items():
-                setattr(self, k, v)
-
-        def to(self, device):
-            import torch as _t
-
-            for k, v in self.__dict__.items():
-                if isinstance(v, _t.Tensor):
-                    setattr(self, k, v.to(device))
-            return self
-
-    sys.modules["torch_geometric.data"].Data = _Data
-
-    class _DataLoader:
-        def __init__(self, dataset, batch_size=1, shuffle=False, **kwargs):
-            self._dataset = list(dataset)
-
-        def __iter__(self):
-            import torch as _t
-
-            for item in self._dataset:
-                for attr in list(vars(item).keys()):
-                    val = getattr(item, attr)
-                    if isinstance(val, (int, float)):
-                        setattr(item, attr, _t.tensor([val]))
-                if not hasattr(item, "num_graphs"):
-                    item.num_graphs = 1
-                if not hasattr(item, "batch"):
-                    n_atoms = item.num_atoms if hasattr(item, "num_atoms") else _t.tensor([0])
-                    if isinstance(n_atoms, _t.Tensor):
-                        n_atoms = int(n_atoms.item())
-                    item.batch = _t.zeros(n_atoms, dtype=_t.long)
-                yield item
-
-        def __len__(self):
-            return len(self._dataset)
-
-    sys.modules["torch_geometric.loader"].DataLoader = _DataLoader
-
-    _make_stub_module("torch_runstats", submodules=["scatter"])
-    import torch as _torch
-
-    def _scatter(src, index, dim_size=None, dim=0, reduce="sum"):
-        if dim_size is None:
-            dim_size = int(index.max()) + 1
-        out = _torch.zeros(dim_size, *src.shape[1:], dtype=src.dtype, device=src.device)
-        if src.dim() == 1:
-            idx = index
-        else:
-            idx = index.unsqueeze(-1).expand_as(src)
-        if reduce == "sum" or reduce == "add":
-            out.scatter_add_(0, idx, src)
-        elif reduce == "mean":
-            out.scatter_add_(0, idx, src)
-            count = _torch.zeros(dim_size, dtype=src.dtype, device=src.device)
-            count.scatter_add_(0, index, _torch.ones(index.shape[0], dtype=src.dtype, device=src.device))
-            count = count.clamp(min=1)
-            if src.dim() > 1:
-                count = count.unsqueeze(-1)
-            out = out / count
-        return out
-
-    sys.modules["torch_runstats.scatter"].scatter = _scatter
-
-    def _scatter_mean(src, index, dim_size=None, dim=0):
-        return _scatter(src, index, dim_size=dim_size, dim=dim, reduce="mean")
-
-    sys.modules["torch_runstats.scatter"].scatter_mean = _scatter_mean
-
-    print("✓ MatterSim dependency stubs applied")
-
-
-# ==============================================================================
-# NequIP patches
-# ==============================================================================
-
-
-def patch_nequip_deps():
-    """
-    Stub heavy dependencies required by NequIP but not needed for inference.
-
-    Stubs: hydra, lightning, pytorch_lightning, torchmetrics, lmdb, matscipy.
-    Patches e3nn and torch.jit for Pyodide compatibility.
-    Sets NEQUIP_NL=ase to use ASE neighbor lists instead of matscipy.
-    """
-    import os
-
-    import torch
-
-    os.environ["NEQUIP_NL"] = "ase"
-
-    _make_stub_module("lightning_utilities", submodules=["core", "core.rank_zero"])
-
-    def _rank_prefixed_message(msg, rank=None):
-        return msg
-
-    def _rank_zero_only(fn):
-        return fn
-
-    sys.modules["lightning_utilities.core.rank_zero"].rank_prefixed_message = _rank_prefixed_message
-    sys.modules["lightning_utilities.core.rank_zero"].rank_zero_only = _rank_zero_only
-
-    _make_stub_module(
-        "lightning",
-        submodules=[
-            "pytorch",
-            "pytorch.utilities",
-            "pytorch.utilities.seed",
-            "pytorch.utilities.warnings",
-            "pytorch.callbacks",
-        ],
-    )
-
-    class _IsolateRng:
-        def __enter__(self):
-            return self
-
-        def __exit__(self, *a):
-            pass
-
-    def _seed_everything(seed=None, workers=False, **kwargs):
-        pass
-
-    sys.modules["lightning.pytorch.utilities.seed"].isolate_rng = lambda: _IsolateRng()
-    sys.modules["lightning.pytorch"].seed_everything = _seed_everything
-
-    class _PossibleUserWarning(UserWarning):
-        pass
-
-    sys.modules["lightning.pytorch.utilities.warnings"].PossibleUserWarning = _PossibleUserWarning
-
-    class _LightningModule(torch.nn.Module):
-        def __init__(self, *a, **k):
-            super().__init__()
-
-        def log(self, *a, **k):
-            pass
-
-    sys.modules["lightning.pytorch"].LightningModule = _LightningModule
-    sys.modules["lightning"].pytorch = sys.modules["lightning.pytorch"]
-    sys.modules["lightning.pytorch.callbacks"].Callback = type("Callback", (), {})
-
-    _make_stub_module("pytorch_lightning", submodules=["utilities", "utilities.seed"])
-    sys.modules["pytorch_lightning.utilities.seed"].isolate_rng = lambda: _IsolateRng()
-
-    _tm = _make_stub_module("torchmetrics")
-
-    class _Metric(torch.nn.Module):
-        def __init__(self, *a, **k):
-            super().__init__()
-
-        def add_state(self, name, default=None, dist_reduce_fx=None, **k):
-            if default is not None:
-                setattr(self, name, default)
-
-    _tm.Metric = _Metric
-    _tm.MeanMetric = _Metric
-
-    if "packaging" not in sys.modules:
-        try:
-            import packaging.version  # noqa: F401
-        except (ImportError, ModuleNotFoundError):
-            _make_stub_module("packaging", submodules=["version"])
-
-            class _Version:
-                def __init__(self, v):
-                    self._v = str(v)
-                    parts = self._v.split(".")
-                    self.major = int(parts[0]) if parts else 0
-                    self.minor = int(parts[1]) if len(parts) > 1 else 0
-
-                def __lt__(self, other):
-                    return (self.major, self.minor) < (other.major, other.minor)
-
-                def __ge__(self, other):
-                    return not self.__lt__(other)
-
-                def __repr__(self):
-                    return f"Version('{self._v}')"
-
-            sys.modules["packaging.version"].Version = _Version
-            sys.modules["packaging.version"].parse = lambda v: _Version(v)
-
-    if "lmdb" not in sys.modules:
-        sys.modules["lmdb"] = types.ModuleType("lmdb")
-
-    if "hydra" not in sys.modules:
-        _make_stub_module(
-            "hydra",
-            submodules=[
-                "core",
-                "core.global_hydra",
-                "utils",
-                "_internal",
-                "_internal.instantiate",
-                "_internal.instantiate._instantiate2",
-            ],
-        )
-
-        def _hydra_instantiate(config, *args, _recursive_=True, **kwargs):
-            import importlib
-
-            if isinstance(config, dict) and "_target_" in config:
-                target = config["_target_"]
-                mod_path, cls_name = target.rsplit(".", 1)
-                mod = importlib.import_module(mod_path)
-                cls = getattr(mod, cls_name)
-                pos_args = list(args) + list(config.get("_args_", []))
-                cfg = {k: v for k, v in config.items() if not k.startswith("_")}
-                if _recursive_:
-                    for k, v in cfg.items():
-                        if isinstance(v, dict) and "_target_" in v:
-                            cfg[k] = _hydra_instantiate(v)
-                        elif isinstance(v, list):
-                            cfg[k] = [
-                                _hydra_instantiate(i) if isinstance(i, dict) and "_target_" in i else i for i in v
-                            ]
-                    pos_args = [
-                        _hydra_instantiate(a) if isinstance(a, dict) and "_target_" in a else a for a in pos_args
-                    ]
-                return cls(*pos_args, **{**cfg, **kwargs})
-            return config
-
-        sys.modules["hydra.utils"].instantiate = _hydra_instantiate
-        sys.modules["hydra"].utils = sys.modules["hydra.utils"]
-        sys.modules["hydra._internal.instantiate._instantiate2"].InstantiationException = type(
-            "InstantiationException", (Exception,), {}
-        )
-
-        def _hydra_get_target(target_str):
-            import importlib
-
-            mod_path, name = target_str.rsplit(".", 1)
-            return getattr(importlib.import_module(mod_path), name)
-
-        sys.modules["hydra.utils"].get_method = _hydra_get_target
-        sys.modules["hydra.utils"].get_class = _hydra_get_target
-
-    if "omegaconf" not in sys.modules:
-        omegaconf_mod = _make_stub_module("omegaconf")
-
-        class _DictConfig(dict):
-            pass
-
-        class _ListConfig(list):
-            pass
-
-        omegaconf_mod.DictConfig = _DictConfig
-        omegaconf_mod.ListConfig = _ListConfig
-        omegaconf_mod.OmegaConf = type(
-            "OmegaConf",
-            (),
-            {
-                "to_container": staticmethod(lambda cfg, **k: dict(cfg) if isinstance(cfg, dict) else cfg),
-                "create": staticmethod(lambda d: _DictConfig(d) if isinstance(d, dict) else d),
-                "register_new_resolver": staticmethod(lambda name, func, **k: None),
-            },
-        )
-
-    if "matscipy" not in sys.modules:
-        _matscipy = types.ModuleType("matscipy")
-        _matscipy.__path__ = []
-        _matscipy.__package__ = "matscipy"
-        _matscipy_neighbours = types.ModuleType("matscipy.neighbours")
-        _matscipy_neighbours.neighbour_list = _matscipy_neighbour_list_compat
-        _matscipy.neighbours = _matscipy_neighbours
-        sys.modules["matscipy"] = _matscipy
-        sys.modules["matscipy.neighbours"] = _matscipy_neighbours
-
-    try:
-        import e3nn
-
-        e3nn._SO3_INITIALIZED = True
-        if hasattr(e3nn, "_OPT_DEFAULTS"):
-            e3nn._OPT_DEFAULTS["jit_mode"] = "eager"
-    except Exception:
-        pass
-
-    if not hasattr(torch.jit, "_original_script"):
-        _orig_script = torch.jit.script
-
-        def _noop_script(obj=None, *a, **k):
-            if obj is not None:
-                return obj
-            return lambda fn: fn
-
-        torch.jit.script = _noop_script
-
-    try:
-        import e3nn.util.jit
-
-        e3nn.util.jit.compile_mode = lambda mode: lambda cls: cls
-    except Exception:
-        pass
-
-    if "tqdm" not in sys.modules:
-        tqdm_mod = _make_stub_module("tqdm", submodules=["auto", "std"])
-
-        def _tqdm_passthrough(iterable=None, *a, **k):
-            return iterable if iterable is not None else iter([])
-
-        tqdm_mod.tqdm = _tqdm_passthrough
-        sys.modules["tqdm.auto"].tqdm = _tqdm_passthrough
-        sys.modules["tqdm.std"].tqdm = _tqdm_passthrough
-
-    print("✓ NequIP dependency stubs applied")
-
-
-def load_nequip_model(checkpoint_path):
-    import torch
-
-    data = torch.load(checkpoint_path, map_location="cpu", weights_only=False)
-
-    import nequip.utils.global_state as _gs
-
-    def _pyodide_set_global_state(allow_tf32=False, warn_on_override=False):
-        if not _gs._GLOBAL_STATE_INITIALIZED:
-            torch.set_default_dtype(torch.float64)
-            try:
-                import e3nn
-
-                e3nn.set_optimization_defaults(
-                    specialized_code=True,
-                    optimize_einsums=True,
-                    jit_script_fx=False,
-                )
-            except Exception:
-                pass
-            _gs._GLOBAL_STATE_INITIALIZED = True
-        _gs._latest_global_config["allow_tf32"] = allow_tf32
-
-    _gs.set_global_state = _pyodide_set_global_state
-    _gs.set_global_state(allow_tf32=False)
-
-    from nequip.model import FullNequIPGNNModel
-
-    model = FullNequIPGNNModel(
-        seed=0,
-        model_dtype="float32",
-        r_max=data["r_max"],
-        type_names=data["type_names"],
-        irreps_edge_sh=data["irreps_edge_sh"],
-        type_embed_num_features=data["type_embed_num_features"],
-        feature_irreps_hidden=data["feature_irreps_hidden"],
-        radial_mlp_depth=data["radial_mlp_depth"],
-        radial_mlp_width=data["radial_mlp_width"],
-        avg_num_neighbors=data["avg_num_neighbors"],
-        per_type_energy_scales=data["per_type_energy_scales"],
-        per_type_energy_shifts=data["per_type_energy_shifts"],
-        polynomial_cutoff_p=data.get("polynomial_cutoff_p", 6),
-    )
-
-    if data.get("has_zbl", False):
-        from nequip.nn.pair_potential import ZBL
-
-        seq_net = model.model.func
-        zbl = ZBL(
-            type_names=data["type_names"],
-            chemical_species=data["type_names"],
-            units="metal",
-            irreps_in=seq_net.irreps_out,
-        )
-        seq_net.insert(name="pair_potential", module=zbl, before="total_energy_sum")
-
-    model.load_state_dict(data["state_dict"], strict=True)
-    model.eval()
-
-    print(f"✓ NequIP model loaded ({sum(p.numel() for p in model.parameters()):,} parameters)")
-    return model
